Add "LoadBalancer" so app will be reachable from outside
- only create load balancer if property spring.cloud.deployer.kubernetes.createLoadBalancer is true

- we must wait for load balancer creation to be complete when deleting service

- updated fabric8 kubernetes-client version

+++ src/main/java/org/springframework/cloud/deployer/spi/kubernetes/DefaultContainerFactory.java
@@ -90,7 +90,7 @@ public class DefaultContainerFactory implements ContainerFactory {
 	/**
 	 * Create a readiness probe for the /health endpoint exposed by each module.
 	 */
-	protected Probe createProbe(Integer externalPort, long timeout, long initialDelay) {
+	protected Probe createProbe(Integer externalPort, int timeout, int initialDelay) {
 		return new ProbeBuilder()
 			.withHttpGet(
 				new HTTPGetActionBuilder()
